chore(profile): moved all profile related core functionality into plugin
Fixes #10300
