+++ mod/profile/elgg-plugin.php
@@ -6,4 +6,19 @@ return [
 	'upgrades' => [
 		AnnotationMigration::class,
 	],
+	'actions' => [
+		'profile/edit' => [],
+		'profile/fields/reset' => [
+			'access' => 'admin',
+		],
+		'profile/fields/add' => [
+			'access' => 'admin',
+		],
+		'profile/fields/delete' => [
+			'access' => 'admin',
+		],
+		'profile/fields/reorder' => [
+			'access' => 'admin',
+		],
+	],
 ];
