+++ mod/profile/start.php
@@ -6,6 +6,7 @@
  */
 
 elgg_register_event_handler('init', 'system', 'profile_init', 1);
+elgg_register_event_handler('init', 'system', '_profile_fields_setup', 10000); // Ensure this runs after other plugins
 
 // Metadata on users needs to be independent
 // outside of init so it happens earlier in boot. See #3316
@@ -22,6 +23,8 @@ function profile_init() {
 	elgg_register_page_handler('profile', 'profile_page_handler');
 
 	elgg_extend_view('elgg.css', 'profile/css');
+	
+	elgg_register_ajax_view('forms/profile/fields/add');
 
 	// allow ECML in parts of the profile
 	elgg_register_plugin_hook_handler('get_views', 'ecml', 'profile_ecml_views_hook');
@@ -31,6 +34,9 @@ function profile_init() {
 	
 	elgg_register_plugin_hook_handler('register', 'menu:topbar', '_profile_topbar_menu');
 	elgg_register_plugin_hook_handler('register', 'menu:title', '_profile_title_menu');
+	elgg_register_plugin_hook_handler('register', 'menu:page', '_profile_admin_page_menu');
+	elgg_register_plugin_hook_handler('register', 'menu:page', '_profile_user_page_menu');
+	elgg_register_plugin_hook_handler('register', 'menu:user_hover', '_profile_user_hover_menu');
 }
 
 /**
@@ -123,19 +129,70 @@ function profile_default_widgets_hook($hook, $type, $return) {
 }
 
 /**
+ * This function loads a set of default fields into the profile, then triggers a hook letting other plugins to edit
+ * add and delete fields.
+ *
+ * Note: This is a secondary system:init call and is run at a super low priority to guarantee that it is called after all
+ * other plugins have initialised.
+ * @access private
+ */
+function _profile_fields_setup() {
+	$profile_defaults =  [
+		'description' => 'longtext',
+		'briefdescription' => 'text',
+		'location' => 'location',
+		'interests' => 'tags',
+		'skills' => 'tags',
+		'contactemail' => 'email',
+		'phone' => 'text',
+		'mobile' => 'text',
+		'website' => 'url',
+		'twitter' => 'text',
+	];
+
+	$loaded_defaults = [];
+	$fieldlist = elgg_get_config('profile_custom_fields');
+	if ($fieldlist || $fieldlist === '0') {
+		$fieldlistarray = explode(',', $fieldlist);
+		foreach ($fieldlistarray as $listitem) {
+			if ($translation = elgg_get_config("admin_defined_profile_{$listitem}")) {
+				$type = elgg_get_config("admin_defined_profile_type_{$listitem}");
+				$loaded_defaults["admin_defined_profile_{$listitem}"] = $type;
+				add_translation(get_current_language(), ["profile:admin_defined_profile_{$listitem}" => $translation]);
+			}
+		}
+	}
+
+	if (count($loaded_defaults)) {
+		elgg_set_config('profile_using_custom', true);
+		$profile_defaults = $loaded_defaults;
+	}
+	
+	$profile_fields = elgg_trigger_plugin_hook('profile:fields', 'profile', null, $profile_defaults);
+	elgg_set_config('profile_fields', $profile_fields);
+
+	// register any tag metadata names
+	foreach ($profile_fields as $name => $type) {
+		if ($type == 'tags' || $type == 'location' || $type == 'tag') {
+			elgg_register_tag_metadata_name($name);
+			// register a tag name translation
+			add_translation(get_current_language(), ["tag_names:$name" => elgg_echo("profile:$name")]);
+		}
+	}
+}
+
+/**
  * Register menu items for the topbar menu
  *
- * @param string $hook
- * @param string $type
- * @param array  $return
- * @param array  $params
+ * @param \Elgg\Hook $hook hook
+ *
  * @return array
  *
  * @access private
  *
  * @since 3.0
  */
-function _profile_topbar_menu($hook, $type, $return, $params) {
+function _profile_topbar_menu(\Elgg\Hook $hook) {
 
 	$viewer = elgg_get_logged_in_user_entity();
 	if (!$viewer) {
@@ -153,6 +210,101 @@ function _profile_topbar_menu($hook, $type, $return, $params) {
 			'class' => 'elgg-border-plain elgg-transition',
 		]),
 		'priority' => 100,
+	]);
+	
+	return $return;
+}
+
+/**
+ * Register menu items for the user hover menu
+ *
+ * @param \Elgg\Hook $hook hook
+ *
+ * @return array
+ *
+ * @access private
+ *
+ * @since 3.0
+ */
+function _profile_user_hover_menu(\Elgg\Hook $hook) {
+
+	if (!elgg_is_logged_in()) {
+		return;
+	}
+	
+	$user = $hook->getEntityParam();
+	if (!($user instanceof \ElggUser) || !$user->canEdit()) {
+		return;
+	}
+	
+	$return = $hook->getValue();
+	$return[] = ElggMenuItem::factory([
+		'name' => 'profile:edit',
+		'text' => elgg_echo('profile:edit'),
+		'icon' => 'address-card',
+		'href' => "profile/$user->username/edit",
+		'section' => (elgg_get_logged_in_user_guid() == $user->guid)? 'action' : 'admin',
+	]);
+	
+	return $return;
+}
+
+/**
+ * Register menu items for the admin page menu
+ *
+ * @param \Elgg\Hook $hook hook
+ *
+ * @return array
+ *
+ * @access private
+ *
+ * @since 3.0
+ */
+function _profile_admin_page_menu(\Elgg\Hook $hook) {
+
+	if (!elgg_in_context('admin') || !elgg_is_admin_logged_in()) {
+		return;
+	}
+	
+	$return = $hook->getValue();
+	
+	$return[] = \ElggMenuItem::factory([
+		'name' => 'configure_utilities:profile_fields',
+		'text' => elgg_echo('admin:configure_utilities:profile_fields'),
+		'href' => 'admin/configure_utilities/profile_fields',
+		'section' => 'configure',
+		'parent_name' => 'configure_utilities',
+	]);
+	
+	return $return;
+}
+
+/**
+ * Register menu items for the page menu
+ *
+ * @param \Elgg\Hook $hook hook
+ *
+ * @return array
+ *
+ * @access private
+ *
+ * @since 3.0
+ */
+function _profile_user_page_menu(\Elgg\Hook $hook) {
+
+	$owner = elgg_get_page_owner_entity();
+	if (!$owner instanceof \ElggUser) {
+		return;
+	}
+	
+	$return = $hook->getValue();
+	
+	$return[] = \ElggMenuItem::factory([
+		'name' => 'edit_profile',
+		'href' => "profile/{$owner->username}/edit",
+		'text' => elgg_echo('profile:edit'),
+		'section' => '1_profile',
+		'contexts' => ['settings'],
 	]);
 	
 	return $return;
