--- views/default/admin/configure_utilities/profile_fields.php
@@ -1,24 +1,0 @@
-<?php
-/**
- * Admin area: edit default profile fields
- */
-
-$add = elgg_view_form('profile/fields/add', ['class' => 'elgg-form-settings'], []);
-$list = elgg_view('admin/configure_utilities/profile_fields/list');
-
-$reset = elgg_view('output/url', [
-	'text' => elgg_echo('reset'),
-	'href' => 'action/profile/fields/reset',
-	'title' => elgg_echo('profile:resetdefault'),
-	'confirm' => elgg_echo('profile:resetdefault:confirm'),
-	'class' => 'elgg-button elgg-button-cancel',
-	'is_trusted' => 'true',
-]);
-
-$body = <<<__HTML
-$add
-$list
-<div class="mtl">$reset</div>
-__HTML;
-
-echo $body;
